Battlefields: Move BF scripts out of game
Port From (https://github.com/TrinityCore/TrinityCore/commit/49523a74a4c28e5ae17f9a8352aa8224b765b7d8)

+++ Source/Game/Arenas/Arena.cs
@@ -77,8 +77,8 @@ namespace Game.Arenas
 
         public override void FillInitialWorldStates(InitWorldStates packet)
         {
-            packet.AddState(ArenaWorldStates.AlivePlayersGreen, (int)GetAlivePlayersCountByTeam(Team.Horde));
-            packet.AddState(ArenaWorldStates.AlivePlayersGold, (int)GetAlivePlayersCountByTeam(Team.Alliance));
+            packet.AddState(ArenaWorldStates.AlivePlayersGreen, GetAlivePlayersCountByTeam(Team.Horde));
+            packet.AddState(ArenaWorldStates.AlivePlayersGold, GetAlivePlayersCountByTeam(Team.Alliance));
         }
 
         void UpdateArenaWorldState()
